Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     printf("enter two numbers: ");
-    scanf("%.2f %.2f" ,&a,&b);
-    printf("sum: %.2f\n" ,a+b);
+    scanf("%d %d" ,&a,&b);
+    printf("sum: %d\n" ,a+b);
     return 0;
 }
